+++ packages/iocuak/src/container/services/api/ContainerModuleServiceApi.ts
@@ -1,5 +1,5 @@
 import { ContainerModuleApi } from '../../modules/api/ContainerModuleApi';
 
-export interface ContainerModuleApiService {
+export interface ContainerModuleServiceApi {
   load(containerModuleApi: ContainerModuleApi): void;
 }