+++ packages/iocuak/src/container/services/api/ContainerServiceApi.ts
@@ -1,0 +1,8 @@
+import { ContainerBindingServiceApi } from './ContainerBindingServiceApi';
+import { ContainerInstanceServiceApi } from './ContainerInstanceServiceApi';
+import { ContainerModuleServiceApi } from './ContainerModuleServiceApi';
+
+export interface ContainerServiceApi
+  extends ContainerBindingServiceApi,
+    ContainerInstanceServiceApi,
+    ContainerModuleServiceApi {}