+++ packages/iocuak/src/container/services/api/ContainerServiceApiImplementation.spec.ts
@@ -15,11 +15,11 @@ import { ContainerModuleApi } from '../../modules/api/ContainerModuleApi';
 import { ContainerBindingService } from '../domain/ContainerBindingService';
 import { ContainerInstanceService } from '../domain/ContainerInstanceService';
 import { ContainerService } from '../domain/ContainerService';
-import { ContainerApiServiceImplementation } from './ContainerApiServiceImplementation';
+import { ContainerServiceApiImplementation } from './ContainerServiceApiImplementation';
 
-describe(ContainerApiServiceImplementation.name, () => {
+describe(ContainerServiceApiImplementation.name, () => {
   let containerServiceMock: jest.Mocked<ContainerService>;
-  let containerApiServiceImplementation: ContainerApiServiceImplementation;
+  let containerServiceApiImplementation: ContainerServiceApiImplementation;
 
   beforeAll(() => {
     containerServiceMock = {
@@ -38,7 +38,7 @@ describe(ContainerApiServiceImplementation.name, () => {
       jest.Mocked<ContainerService>
     > as jest.Mocked<ContainerService>;
 
-    containerApiServiceImplementation = new ContainerApiServiceImplementation(
+    containerServiceApiImplementation = new ContainerServiceApiImplementation(
       containerServiceMock,
     );
   });
@@ -58,7 +58,7 @@ describe(ContainerApiServiceImplementation.name, () => {
         ).mockReturnValueOnce(undefined);
 
         try {
-          containerApiServiceImplementation.bind(typeFixture);
+          containerServiceApiImplementation.bind(typeFixture);
         } catch (error: unknown) {
           result = error;
         }
@@ -100,7 +100,7 @@ describe(ContainerApiServiceImplementation.name, () => {
           >
         ).mockReturnValueOnce(bindingFixture);
 
-        containerApiServiceImplementation.bind(typeFixture);
+        containerServiceApiImplementation.bind(typeFixture);
       });
 
       afterAll(() => {
@@ -127,7 +127,7 @@ describe(ContainerApiServiceImplementation.name, () => {
           foo: 'bar',
         };
 
-        containerApiServiceImplementation.bindToValue(
+        containerServiceApiImplementation.bindToValue(
           serviceIdFixture,
           instanceFixture,
         );
@@ -168,7 +168,7 @@ describe(ContainerApiServiceImplementation.name, () => {
           instanceFixture,
         );
 
-        result = containerApiServiceImplementation.get(serviceIdFixture);
+        result = containerServiceApiImplementation.get(serviceIdFixture);
       });
 
       afterAll(() => {
@@ -218,7 +218,7 @@ describe(ContainerApiServiceImplementation.name, () => {
           convertBindingToBindingApi as jest.Mock<BindingApi, [Binding]>
         ).mockReturnValueOnce(bindingApiFixture);
 
-        result = containerApiServiceImplementation.getAllBindinds();
+        result = containerServiceApiImplementation.getAllBindinds();
       });
 
       afterAll(() => {
@@ -250,7 +250,7 @@ describe(ContainerApiServiceImplementation.name, () => {
           load: jest.fn(),
         };
 
-        containerApiServiceImplementation.load(containerModuleApiMock);
+        containerServiceApiImplementation.load(containerModuleApiMock);
       });
 
       afterAll(() => {
@@ -270,7 +270,7 @@ describe(ContainerApiServiceImplementation.name, () => {
       beforeAll(() => {
         serviceIdFixture = 'service-id';
 
-        containerApiServiceImplementation.unbind(serviceIdFixture);
+        containerServiceApiImplementation.unbind(serviceIdFixture);
       });
 
       afterAll(() => {