+++ packages/iocuak/src/container/services/api/ContainerServiceApiImplementation.ts
@@ -8,9 +8,9 @@ import { ValueBinding } from '../../../metadata/models/domain/ValueBinding';
 import { convertBindingToBindingApi } from '../../../metadata/utils/api/convertBindingToBindingApi';
 import { ContainerModuleApi } from '../../modules/api/ContainerModuleApi';
 import { ContainerService } from '../domain/ContainerService';
-import { ContainerApiService } from './ContainerApiService';
+import { ContainerServiceApi } from './ContainerServiceApi';
 
-export class ContainerApiServiceImplementation implements ContainerApiService {
+export class ContainerServiceApiImplementation implements ContainerServiceApi {
   protected readonly _containerService: ContainerService;
 
   constructor(containerService: ContainerService) {