+++ packages/iocuak/src/index.ts
@@ -2,7 +2,7 @@ import { Newable } from './common/models/domain/Newable';
 import { ServiceId } from './common/models/domain/ServiceId';
 import { ContainerApi } from './container/modules/api/ContainerApi';
 import { ContainerModuleApi } from './container/modules/api/ContainerModuleApi';
-import { ContainerApiService } from './container/services/api/ContainerApiService';
+import { ContainerServiceApi } from './container/services/api/ContainerServiceApi';
 import { inject } from './metadata/decorators/inject';
 import { injectable } from './metadata/decorators/injectable';
 import { injectFrom } from './metadata/decorators/injectFrom';
@@ -15,7 +15,7 @@ import { TaskScopeApi } from './task/models/api/TaskScopeApi';
 
 export type {
   BindingApi,
-  ContainerApiService,
+  ContainerServiceApi,
   ContainerModuleApi,
   InjectableOptionsApi,
   Newable,