+++ packages/iocuak/src/metadata/services/api/MetadataServiceApi.ts
@@ -2,7 +2,7 @@ import { Newable } from '../../../common/models/domain/Newable';
 import { ClassMetadataApi } from '../../models/api/ClassMetadataApi';
 import { TypeBindingApi } from '../../models/api/TypeBindingApi';
 
-export interface MetadataApiService {
+export interface MetadataServiceApi {
   getBindingMetadata<TInstance, TArgs extends unknown[]>(
     type: Newable<TInstance, TArgs>,
   ): TypeBindingApi<TInstance, TArgs> | undefined;