+++ packages/iocuak/CHANGELOG.md
@@ -28,8 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Updated `ContainerApi` with `build` and `createChild` methods.
-- `ContainerApiService` interface is now exposed.
-- Updated `ContainerApiService` with `getAllBindinds` method
+- `ContainerServiceApi` interface is now exposed.
 - [BC]: Updated `ContainerApi` constructor to be private.
 - [BC]: Renamed `BindingApi` to `InjectableOptionsApi`.
 - [BC]: Renamed `TaskScope` to `TaskScopeApi`.